Moved rayon impls in a separated module
