+++ src/lib.rs
@@ -49,13 +49,6 @@ use crate::maybe_uninit::MaybeUninit;
 #[cfg(feature="serde")]
 use serde::{Serialize, Deserialize, Serializer, Deserializer};
 
-#[cfg(feature="rayon")]
-use rayon::iter::{
-    plumbing::{bridge, Consumer, Producer, ProducerCallback, UnindexedConsumer},
-    FromParallelIterator, IndexedParallelIterator, IntoParallelIterator, ParallelExtend,
-    ParallelIterator,
-};
-
 mod array;
 mod array_string;
 mod char;
@@ -66,6 +59,10 @@ use crate::array::Index;
 pub use crate::array_string::ArrayString;
 pub use crate::errors::CapacityError;
 
+#[cfg(feature="rayon")]
+mod rayon_impls;
+#[cfg(feature="rayon")]
+pub use rayon_impls::IntoParIter;
 
 /// A vector with a fixed capacity.
 ///
@@ -1201,223 +1198,5 @@ impl<'de, T: Deserialize<'de>, A: Array<Item=T>> Deserialize<'de> for ArrayVec<A
         }
 
         deserializer.deserialize_seq(ArrayVecVisitor::<T, A>(PhantomData))
-    }
-}
-
-// Adapted from `rayon/srs/vec.rs`
-
-#[cfg(feature = "rayon")]
-/// Parallel iterator that moves out of an `ArrayVec`.
-#[derive(Debug, Clone)]
-pub struct IntoParIter<T, A: Array<Item = T>> {
-    vec: ArrayVec<A>,
-}
-
-#[cfg(feature = "rayon")]
-impl<A> IntoParallelIterator for ArrayVec<A>
-where
-    A: Array + Send,
-    A::Item: Send,
-    A::Index: Send,
-{
-    type Item = A::Item;
-    type Iter = IntoParIter<A::Item, A>;
-
-    fn into_par_iter(self) -> Self::Iter {
-        IntoParIter { vec: self }
-    }
-}
-
-#[cfg(feature = "rayon")]
-impl<A> ParallelIterator for IntoParIter<A::Item, A>
-where
-    A: Array + Send,
-    A::Item: Send,
-    A::Index: Send,
-{
-    type Item = A::Item;
-
-    fn drive_unindexed<C>(self, consumer: C) -> C::Result
-    where
-        C: UnindexedConsumer<Self::Item>,
-    {
-        bridge(self, consumer)
-    }
-
-    fn opt_len(&self) -> Option<usize> {
-        Some(self.len())
-    }
-}
-
-#[cfg(feature = "rayon")]
-impl<A> IndexedParallelIterator for IntoParIter<A::Item, A>
-where
-    A: Array + Send,
-    A::Item: Send,
-    A::Index: Send,
-{
-    fn drive<C>(self, consumer: C) -> C::Result
-    where
-        C: Consumer<Self::Item>,
-    {
-        bridge(self, consumer)
-    }
-
-    fn len(&self) -> usize {
-        self.vec.len()
-    }
-
-    fn with_producer<CB>(mut self, callback: CB) -> CB::Output
-    where
-        CB: ProducerCallback<Self::Item>,
-    {
-        // The producer will move or drop each item from its slice, effectively taking ownership of
-        // them.  When we're done, the vector only needs to free its buffer.
-        unsafe {
-            // Make the `ArrayVec` forget about the actual items.
-            let len = self.vec.len();
-            self.vec.set_len(0);
-
-            // Get a correct borrow, then extend it to the original length.
-            let mut slice = self.vec.as_mut_slice();
-            slice = std::slice::from_raw_parts_mut(slice.as_mut_ptr(), len);
-
-            callback.callback(ArrayVecProducer { slice })
-        }
-    }
-}
-
-/// ////////////////////////////////////////////////////////////////////////
-
-#[cfg(feature = "rayon")]
-struct ArrayVecProducer<'data, T: Send> {
-    slice: &'data mut [T],
-}
-
-#[cfg(feature = "rayon")]
-impl<'data, T: 'data + Send> Producer for ArrayVecProducer<'data, T> {
-    type Item = T;
-    type IntoIter = SliceDrain<'data, T>;
-
-    fn into_iter(mut self) -> Self::IntoIter {
-        // replace the slice so we don't drop it twice
-        let slice = std::mem::replace(&mut self.slice, &mut []);
-        SliceDrain {
-            iter: slice.iter_mut(),
-        }
-    }
-
-    fn split_at(mut self, index: usize) -> (Self, Self) {
-        // replace the slice so we don't drop it twice
-        let slice = std::mem::replace(&mut self.slice, &mut []);
-        let (left, right) = slice.split_at_mut(index);
-        (
-            ArrayVecProducer { slice: left },
-            ArrayVecProducer { slice: right },
-        )
-    }
-}
-
-#[cfg(feature = "rayon")]
-impl<'data, T: 'data + Send> Drop for ArrayVecProducer<'data, T> {
-    fn drop(&mut self) {
-        SliceDrain {
-            iter: self.slice.iter_mut(),
-        };
-    }
-}
-
-/// ////////////////////////////////////////////////////////////////////////
-
-// like std::vec::Drain, without updating a source Vec
-#[cfg(feature = "rayon")]
-struct SliceDrain<'data, T> {
-    iter: std::slice::IterMut<'data, T>,
-}
-
-#[cfg(feature = "rayon")]
-impl<'data, T: 'data> Iterator for SliceDrain<'data, T> {
-    type Item = T;
-
-    fn next(&mut self) -> Option<T> {
-        let ptr = self.iter.next()?;
-        Some(unsafe { std::ptr::read(ptr) })
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let len = self.len();
-        (len, Some(len))
-    }
-}
-
-#[cfg(feature = "rayon")]
-impl<'data, T: 'data> DoubleEndedIterator for SliceDrain<'data, T> {
-    fn next_back(&mut self) -> Option<Self::Item> {
-        let ptr = self.iter.next_back()?;
-        Some(unsafe { std::ptr::read(ptr) })
-    }
-}
-
-#[cfg(feature = "rayon")]
-impl<'data, T: 'data> ExactSizeIterator for SliceDrain<'data, T> {
-    fn len(&self) -> usize {
-        self.iter.len()
-    }
-}
-
-#[cfg(feature = "rayon")]
-impl<'data, T: 'data> Drop for SliceDrain<'data, T> {
-    fn drop(&mut self) {
-        for ptr in &mut self.iter {
-            unsafe {
-                std::ptr::drop_in_place(ptr);
-            }
-        }
-    }
-}
-
-#[cfg(feature = "rayon")]
-impl<A> FromParallelIterator<A::Item> for ArrayVec<A>
-where
-    A: Array + Send,
-    A::Item: Send,
-    A::Index: Send,
-{
-    fn from_par_iter<I>(par_iter: I) -> Self
-    where
-        I: IntoParallelIterator<Item = A::Item>,
-    {
-        let mut arrayvec = Self::new();
-        arrayvec.par_extend(par_iter);
-        arrayvec
-    }
-}
-
-#[cfg(feature = "rayon")]
-impl<A> ParallelExtend<A::Item> for ArrayVec<A>
-where
-    A: Array + Send,
-    A::Item: Send,
-    A::Index: Send,
-{
-    fn par_extend<I>(&mut self, par_iter: I)
-    where
-        I: IntoParallelIterator<Item = A::Item>,
-    {
-        self.extend(
-            par_iter
-                .into_par_iter()
-                .fold(
-                    || Self::new(),
-                    |mut arrayvec, element| {
-                        let _ = arrayvec.try_push(element);
-                        arrayvec
-                    },
-                )
-                .reduce(Self::new, |mut arrayvec1, arrayvec2| {
-                    arrayvec1.extend(arrayvec2);
-                    arrayvec1
-                }),
-        )
     }
 }
